Change to make envvars a list of environment variables to pass
Previously, this was a list of var=value settings.  This change
makes it so you specify the list of variable to pass from the
environment.  It will iterate through the names and add the
name={value from environment} to the azure deploy line.  These
same names must be set in the environment section.